ACM-23769 Application--organize folders and files
