+++ frontend/src/routes/Applications/Applications.tsx
@@ -4,10 +4,10 @@ import ApplicationsDetailsPage from './ApplicationDetails/ApplicationDetails'
 import ApplicationsPage from './ApplicationsPage'
 import Overview from './Overview'
 import AdvancedConfiguration from './AdvancedConfiguration'
-import { CreateApplicationArgo } from './CreateApplication/CreateApplicationArgo'
-import { EditArgoApplicationSet } from './CreateApplication/EditArgoApplicationSet'
-import { CreateApplicationArgoPullModel } from './CreateApplication/CreateApplicationArgoPullModel'
-import CreateSubscriptionApplicationPage from './SubscriptionApplication'
+import { CreateApplicationArgo } from './CreateArgoApplication/CreateApplicationArgo'
+import { EditArgoApplicationSet } from './CreateArgoApplication/EditArgoApplicationSet'
+import { CreateApplicationArgoPullModel } from './CreateArgoApplication/CreateApplicationArgoPullModel'
+import CreateSubscriptionApplicationPage from './CreateSubscriptionApplication/SubscriptionApplication'
 import { MatchType, NavigationPath, SubRoutesRedirect, createRoutePathFunction } from '../../NavigationPath'
 import { ApplicationOverviewPageContent } from './ApplicationDetails/ApplicationOverview/ApplicationOverview'
 import { ApplicationTopologyPageContent } from './ApplicationDetails/ApplicationTopology/ApplicationTopology'
