--- frontend/src/routes/Applications/CreateApplication/Subscription/style.css
@@ -1,27 +1,0 @@
-.temptifly {
-  margin: 0;
-}
-
-.temptifly.showEditor,
-.temptifly.showWizard {
-  height: 100%;
-  margin: 0;
-  padding: 0;
-}
-
-.temptifly .creation-view-split-container .creation-view-controls-container,
-.temptifly .creation-view-controls-container {
-  padding: 0;
-}
-
-.temptifly .creation-view-split-container.showEditor .creation-view-controls-container,
-.temptifly.showWizard .creation-view-controls-container {
-  padding: 0;
-}
-
-.portal-controls {
-  display: flex;
-}
-.portal-controls > div {
-  margin-left: 16px;
-}
