+++ frontend/src/routes/Applications/CreateArgoApplication/CreateApplicationArgo.test.tsx
