+++ frontend/src/routes/Applications/CreateArgoApplication/actions/index.tsx
@@ -1,0 +1,2 @@
+/* Copyright Contributors to the Open Cluster Management project */
+export const argoAppSetQueryString = '?' + 'apiVersion=applicationset.argoproj.io'
