+++ frontend/src/routes/Applications/CreateArgoApplication/schema.json
