+++ frontend/src/routes/Applications/CreateSubscriptionApplication/actions/index.tsx
@@ -1,4 +1,2 @@
 /* Copyright Contributors to the Open Cluster Management project */
-export const argoAppSetQueryString = '?' + 'apiVersion=applicationset.argoproj.io'
-
 export const subscriptionAppQueryString = '?' + 'apiVersion=application.app.k8s.io'
