+++ frontend/src/routes/Applications/CreateSubscriptionApplication/components/ClusterSelector/index.test.tsx
