+++ frontend/src/routes/Applications/CreateSubscriptionApplication/components/style.css
