+++ frontend/src/routes/Applications/CreateSubscriptionApplication/style.css
