+++ frontend/src/routes/Applications/CreateSubscriptionApplication/templates/template.hbs
