+++ frontend/src/routes/Applications/CreateSubscriptionApplication/templates/templateGit.hbs
