+++ frontend/src/routes/Applications/CreateSubscriptionApplication/templates/templateObjectStore.hbs
