+++ frontend/src/routes/Applications/CreateSubscriptionApplication/templates/templatePlacement.hbs
