+++ frontend/src/routes/Applications/CreateSubscriptionApplication/transformers/transform-data-to-resources.tsx
